service/requirements.txt
pytest
pytest-mock

influxdb-client
pyyaml
requests